Rewrite Charts, step (29/30): remove the code using Charts
Ignore-this: b01b22e877448d79398c0c23d8468813

I've just compared the results of drawing a few test eventlogs with Charts
and with our own histogram code and the numbers agree, but the graphs
are different when an interval is selected. Charts change both scales
and the number of bars when the data changes, while our code keeps
the scales unchanged, similarly as the other spark graphs do. TODO: rescale
when the "zoom to fit" button is pressed or when an appropriate options
is selected.

darcs-hash:20111107134246-9f033-f280683831b1403d2be519cf5371bd4400b6f556