Actual implementation of hint
We want the gray hints in the source tag view to follow the selection
in the Core window. Right now this is less useful than it could be,
as it obviously only works for tags that are actually *present*.
Therefore anything that has zero cost associated with it in the Core
view will not have a hint, even though we actually know its source
annotations.