Move allocation cost where it improves consistency
This works around an unfortunate interaction with GHC's handling of
allocation - the return pointer we get as sampling output will actually
often lead us "back" in the execution, as GHC often reasons that going
a bit back in execution is preferable to having to do more heap checks.

Unfortunately, this can lead us to situations where we positively can't
decide what allocation site was actually responsible for triggering
a block request. This code now more or less awkwardly tries to move
cost as far "inwards" as possible until it can't decide anymore what
direction to go. This gracefully resolves about 80% of the cases, and in
the remaining 20% it possibly minimises the confusion.

Not very satisfactory, though.