Sample maps
The goal here is to quickly get an overview on where in the profile
certain IP ranges appear. We optimize mainly two aspects - first, we
build a map with all samples of a certain type that can be found in the
profile (this will probably be handy for other things too). Second, we
make extra sure that the sampling loop doesn't do double-checks or ends
up using IP ranges where higher-priority ones exist. The result gets
converted into a time-map for fast overview generation.