Timeline hints
This highlights an area on the time-line corresponding approximately
to the proportion of time spent for the selected profiling noun.

Note that we do this exactly for the activity graph - which is quite wrong
if the profiling verb selected happens to be heap allocation or heap data.
We need heap the actual allocation / heap data profile graphs to draw our
hints over. Yet another TODO that should be right around the corner now...